[Application] Split Http and Cli from Entry\App.
